feat: add executeFile and analyzeFile functionality with routing in code controllers and routes
